Refactors StoreState::get and get_mut methods
Changes the implementation of the `get` and `get_mut` methods in `StoreState` to use `as_ref` and `as_mut` for pointer dereferencing. This improves safety by preventing direct raw pointer access and potentially undefined behavior. Adds `inline` attribute to those methods to suggest inlining for potential performance benefits. Also adds a test case to demonstrate the use of StoreState with an invalid lifetime.